--- a/src/array/data_chunk_builder.rs
+++ b/src/array/data_chunk_builder.rs
@@ -1,7 +1,5 @@
 use std::iter::IntoIterator;
 
-use itertools::Itertools;
-
 use super::{ArrayBuilderImpl, DataChunk};
 use crate::types::{ConvertError, DataType, DataValue};
 
@@ -37,7 +35,7 @@ impl DataChunkBuilder {
     pub fn push_row(&mut self, row: impl IntoIterator<Item = DataValue>) -> Option<DataChunk> {
         self.array_builders
             .iter_mut()
-            .zip_eq(row)
+            .zip(row)
             .for_each(|(builder, v)| builder.push(&v));
         self.size += 1;
         if self.size == self.capacity {
@@ -60,7 +58,7 @@ impl DataChunkBuilder {
         &mut self,
         row: impl IntoIterator<Item = &'a str>,
     ) -> Result<Option<DataChunk>, ConvertError> {
-        for (builder, r) in self.array_builders.iter_mut().zip_eq(row) {
+        for (builder, r) in self.array_builders.iter_mut().zip(row) {
             builder.push_str(r)?
         }